requirements.txt
streamlit
pypdf
sentence-transformers
chromadb
requests
python-dotenv